commit 5
0. env
# status 작성하지 않음

- change -
호환되는 파이썬 버전 작성됨

- need -
None

1. data_loader.py 0.0.0 > 0.0.1
status: finish

- change -
 코드 최적화

 - need -
 None

2. data_normalizers.py  0.0.1 > 0.0.2
status: finish

- change -
 코드 최적화

 - need -
 None

3. index.py
 status: end

- change -
 None

 - need -
 None

4. tokenizer.py 0.0.0
 status: active

- change -
 1. __h2hcj() 함수에서 기존 jamo가 존재하지 않는 위치는 ""를 추가해 3길이를 맞추었으나 이것을 padding 토큰으로 대체
 2. 메서드 사전 정의 (zero vector, one hot vector, encode)
 3. 토큰화, 및 역토큰화 메서드 생성
 4. 코드 최적화

 - need -
 1. 정의된 메서드 개발 및 정상작동 테스트 필요.

5. corpus > type.py 0.0.0
 status: active

- change -
 1. Corpus 타입에 해당하는 데이터 범위를 여유롭게 조정(list[str] > list)
 2. 코드 최적화

 - need -
 None

6. tokenizer > type.py 0.0.0
 status: active

- change -
 1. Jamo에 대한 문자열일때, 수치화 되었을 때의 변수 타입 정의.
 2. DefaultSpecialToken 프로퍼티 이름 변경됨
 3. specical token에 대한 타입 상세 분활됨.

 - need -
 None

7.test > tokenizer > vocab.py 0.0.0 > 0.0.1
 status: finish

- change -
 코드 최적화
 변경점 x

 - need -
 None

 8. tokenizer> jamo > vocab.py 0.0.0 > 0.0.1
  status: finish

- change -
 1. 코드 최적화
 2. 초기 인스턴스 생성 시 special token 적용하는 로직 구성.
 3. JamoTokenizer에 있었던 맴버변수인 special_tokens을 Vocab으로 이동함.

 - need -
 None

+++ tokenizer/jamo/tokenizer.py
@@ -1,22 +1,27 @@
+from typing_extensions import Unpack
+
 import json
 import os
-from typing import Optional
+from typing import Optional, Union
 
-from jamo import j2hcj, h2j, j2h, is_jamo, get_jamo_class, hcj_to_jamo
+import numpy as np
+from jamo import j2hcj, h2j, j2h
 
 from corpus.data_normalizers import DataNormalizers
-from corpus.type import JamoEncodingType, TypePath, NormalizationMethod
+from corpus.type import JamoEncodingType, Path, NormalizerOption
 from tokenizer.jamo.vocab import Vocab
-from tokenizer.type import DefaultSpecialToken, JamoList, Jamo, SpecialToken, SpecialTokenSet
+from tokenizer.type import JamoMatrix, \
+    JamoMatrixDataset, CustomSpecialToken, DefaultSpecialToken, JamoCharArray, JamoCharArrayDataset
 from util import Debug
 
 
-class LoadVocab():
-    vocab_path: TypePath
+class LoadVocab:
+    vocab_path: Path
     debug: Debug
-    def __init__(self, _vocab_path: Optional[TypePath] = None):
+    def __init__(self, _vocab_path: Optional[Path] = None):
         self.debug = Debug(*eval(os.environ.get('DEBUG_OPTION')))
         self.vocab_path = _vocab_path
+
 
     def __check_vocab_exist(self) -> bool:
         r"""path에 vocab이 존재할 시 true를 리턴하며 오류 발생 또는 아닐 시 출력값과 False를 리턴합니다."""
@@ -25,7 +30,7 @@ class LoadVocab():
         else:
             return False
 
-    def load_file(self) -> Vocab:
+    def load_file(self) -> dict:
         r"""vocab을 load하여 self.vocab에 저장합니다."""
         # https://www.geeksforgeeks.org/convert-list-of-tuples-to-json-python/
 
@@ -33,7 +38,7 @@ class LoadVocab():
 
         if self.__check_vocab_exist() is False:
             self.debug.opt_debug_print(f"vocab is not exist in next location ( this message is not error )\ninput path: {self.vocab_path}")
-            return Vocab({})
+            return {}
 
         with open(self.vocab_path, "r", encoding="utf-8") as file:
             if self.vocab_path[-3:] == "txt":
@@ -45,29 +50,28 @@ class LoadVocab():
                 self.debug.unexpected_error(
                     f"vocab must be .txt or .json file.\ninput path: {self.vocab_path}")
 
-            return Vocab(_key_vocab)
+            return _key_vocab
 
 
-class JamoTokenizer():
-    vocab_path: TypePath
+class JamoTokenizer:
+    vocab_path: Path
     sentence_length: int
     encoding_type: JamoEncodingType
-    special_token: SpecialTokenSet
     truncation: bool = False
     padding: bool = False
-    normalizer_queue: list[NormalizationMethod]
+    normalizer_queue: list[NormalizerOption]
     r"""입력된 순서대로 정규화를 진행하며 각 과정은 고유한 특성(중복x)을 갖는다. first input first out"""
 
     vocab: Vocab
     dataNormalizer: DataNormalizers
     debug: Debug
     def __init__(self,
-                 vocab_path: TypePath = "./vocab.json",
+                 vocab_path: Path = "./vocab.json",
                  sentence_length: int = 200,
                  padding: bool = False,
                  truncation: bool = False,
-                 special_token: list[SpecialToken] = None,
-                 normalizer: list[NormalizationMethod] = None,
+                 special_token: list[CustomSpecialToken] = None,
+                 normalizer: list[Unpack[NormalizerOption]] = None,
                  encoding_type: JamoEncodingType = JamoEncodingType.JAMO_VECTOR,
                  ):
         self.vocab_path = vocab_path
@@ -75,35 +79,32 @@ class JamoTokenizer():
 
         self.padding = padding
         self.truncation = truncation
-        self.special_token = [property for property in DefaultSpecialToken ]
-        if special_token is not None:
-            for token in special_token:
-                self.special_token.append(token)
-        
+
+        self.vocab = Vocab(LoadVocab(vocab_path).load_file(), special_token)
+
         self.normalizer_queue = normalizer
         self.encoding_type = encoding_type
 
         self.dataNormalizer = DataNormalizers()
-        self.vocab = LoadVocab(vocab_path).load_file()
         self.debug = Debug(*eval(os.environ.get('DEBUG_OPTION')))
 
-    def __h2hcj(self, content) -> list[Jamo]:
+    def __h2hcj(self, content) -> JamoCharArray:
         r"""
         한글 문장 > h2j(자소 단위 정규화) > j2hcj(호환 자모 정규화)
         """
-        result = []
+        result: list = []
         for char in content:
-            char = list(j2hcj(h2j(char)))
+            char = np.array(j2hcj(h2j(char)))
             if char.__len__() < 3:
                 for i in range(3 - char.__len__()):
-                    char.append("")
-            result.append(Jamo(char))
-        return result
+                    np.append(char, self.vocab.get(DefaultSpecialToken._member_names_[0])) # padding은 0번째에 위치함
+            np.append(result, char)
+        return JamoCharArray(result)
 
-    def __jamo_string_2_h(self, jamo_str: list[Jamo]):
-        r"""(3, vocab_length) shape 형태 백터 한글 변형"""
-        result = ""
-        for char_jamo in jamo_str:
+    def __jamo_string_2_h(self, jamo_array: JamoCharArray) -> str:
+        r"""( 3, jamo ) shape 형태 백터 한글 변형"""
+        result:str = ""
+        for char_jamo in jamo_array:
             print(*char_jamo)
             if char_jamo[0].__len__() == 1 and char_jamo[1].__len__() == 1: # 초성 중성이 존재할 시에만 j2h 메서드 사용
                 result += j2h(*char_jamo)
@@ -124,24 +125,49 @@ class JamoTokenizer():
         self.debug.opt_debug_print("successfully added")
         self.debug.opt_debug_print(self.vocab)
 
-        with open("./tokenizer/data/vocab.json", encoding="utf-8", mode="w") as f:
-            f.write(json.dumps(self.vocab, ensure_ascii= False))
+        if os.path.exists(self.vocab_path):
+            with open(self.vocab_path, encoding="utf-8", mode="w") as f:
+                f.write(json.dumps(self.vocab, ensure_ascii= False))
+        else:
+            with open("./tokenizer/data/vocab.json", encoding="utf-8", mode="w") as f:
+                f.write(json.dumps(self.vocab, ensure_ascii= False))
 
-    def __zero_sparse_vector(self):
-        ... 
 
-    def tokenize(self, sentence: list[str]) -> JamoList:
-        x = self.dataNormalizer.filtering_normalize(sentence=sentence)
-        x = self.__h2hcj(x)
-        return x
+    def __encoding_one_hot_vector(self):
+        ...
+    def __create_zero_vector(self):
+        ...
+
+
+    def tokenize(self, sentence: Union[Optional[str], list[str]]) -> Union[Optional[JamoCharArray], JamoCharArrayDataset]:
+        if isinstance(sentence, JamoCharArrayDataset):
+            self.dataNormalizer.set_corpus(sentence)
+            sentence = self.dataNormalizer.compute_normalize()
+        else:
+            sentence = self.dataNormalizer.filtering_normalize(str(sentence))
+        sentence = self.__h2hcj(sentence)
+        return sentence
     
-    # def combinate()
+    def merge_token(self, sentence: Union[Optional[JamoCharArrayDataset], JamoCharArray]):
+        if isinstance(sentence, JamoCharArray):
+            sentence = self.__jamo_string_2_h(sentence)
+        else:
+            for i, single_sentence in enumerate(sentence):
+                sentence[i] = self.__jamo_string_2_h(JamoCharArray(single_sentence))
+        return sentence
 
-    def encode(self, sentence: str):
-        unit_set = set(sentence)
-        unit_key_pair = []
+    def encode(self, sentence: Union[Optional[JamoMatrix],JamoMatrixDataset]):
+        if isinstance(sentence, JamoMatrix):
+            ...
+        elif isinstance(sentence, JamoMatrixDataset):
+            ...
+        else:
+            self.debug.unexpected_error("sentence argument must be JamoVector or JamoVectorDataset. you must use method \"JamoTokenizer.tokenize\" before sentence encoding.")
 
-        for word in unit_set:
-            self.vocab.token_2_key(word)
+        # unit_set = set()
+        # unit_key_pair = []
+        #
+        # for word in unit_set:
+        #     self.vocab.token_2_key(word)
